requirements.txt
PyYAML>=3
torch>=1.0
click>=6.7
typing>=3.6
mayavi>=4.6
numpy>=1.15
Pillow>=5.2
path.py>=11.5
torchvision>=0.2
tensorboardX>=1.6
opencv_python>=3.4
